First add of resurrection
